requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
tensorflow==2.16.1
aiofiles==23.2.1
email-validator==2.2.0
